Adds basic routing setup
