+++ package.json
@@ -1,21 +1,19 @@
 {
   "name": "find-teso-geek",
   "version": "0.0.0",
-  "description": "",
-  "main": "index.js",
   "scripts": {
     "prebuild": "rm -rf build",
     "build": "webpack --config=webpack.prod.config.js",
-    "develop": "webpack-dev-server --config webpack.dev.config.js --open"
+    "develop": "webpack-dev-server --config webpack.dev.config.js --open --color"
   },
   "author": "Mindaugas Jačionis",
-  "license": "ISC",
   "dependencies": {
     "idx": "^2.5.4",
     "react": "^16.8.3",
     "react-dom": "^16.8.3",
     "react-redux": "^6.0.1",
     "react-router": "^4.3.1",
+    "react-router-dom": "^4.3.1",
     "react-svg-inline": "^2.1.1",
     "redux": "^4.0.1",
     "reset-css": "^4.0.1",
