+++ src/app/pages/Finish.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import { Link } from 'react-router-dom';
+
+function Finish() {
+  return <Link to="/start">Restart The Game</Link>;
+}
+
+export default Finish;
