+++ src/app/pages/Start.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import { Link } from 'react-router-dom';
+
+function Start() {
+  return <Link to="/game">Start the Game</Link>;
+}
+
+export default Start;
