+++ src/app/pages/index.js
@@ -1,0 +1,3 @@
+export { default as Game } from './Game';
+export { default as Start } from './Start';
+export { default as Finish } from './Finish';
