+++ webpack.dev.config.js
@@ -1,11 +1,12 @@
 const merge = require('webpack-merge');
+const path = require('path');
 
 const { paths } = require('./webpack.constants');
 const baseConfig = require('./webpack.base.config');
 
 module.exports = merge.smart(baseConfig, {
   devServer: {
-    contentBase: paths.BUILD,
+    historyApiFallback: true,
     port: '4040',
   },
   devtool: 'inline-source-map',
